For your own sanity never use "List view" in widgets
After spending almost 3 days trying to understand said system I created my own custom list with blackja... easier to use and configure.
The hardest parts of the load system are almost finished.
While I may not be the fastest programmer, I learned a lot about widgets, and this could help me in future projects, which is great.
I know that nobody will read this, but if you do, have a nice day.