Slot savesystem completed
Now it is possible to save and load on multiple slots, each slot with a thumbnail and some save data.
The user is capable to copy and paste saves, change the names or delete them in the folder with no problem(Take into consideration that there is no "get all savegames" function).
The slot limit is 4096 (more than that will increase the load time).