Fixing an internal unreal engine problem in the ramapping system
Problem: the function to return the keys assigned to an action/axis does not return it with a constant index, making it hard to set up keys due that a restart of the game will return they keys assingned with a different index.
This makes impossible the task of assigning multiple bindings to the same action unless doing a complicated walkaround, which is the thing I will try to do.